+++ actionview/lib/action_view/helpers/tag_helper.rb
@@ -32,11 +32,13 @@ module ActionView
 
       ARIA_PREFIXES = ["aria", :aria].to_set.freeze
       DATA_PREFIXES = ["data", :data].to_set.freeze
+      CLASS_PREFIXES = ["class", :class].to_set.freeze
 
       TAG_TYPES = {}
       TAG_TYPES.merge! BOOLEAN_ATTRIBUTES.index_with(:boolean)
       TAG_TYPES.merge! DATA_PREFIXES.index_with(:data)
       TAG_TYPES.merge! ARIA_PREFIXES.index_with(:aria)
+      TAG_TYPES.merge! CLASS_PREFIXES.index_with(:class)
       TAG_TYPES.freeze
 
       PRE_CONTENT_STRINGS             = Hash.new { "" }
@@ -240,7 +242,7 @@ module ActionView
             next if key.blank?
 
             type = TAG_TYPES[key]
-            if type == :data && value.is_a?(Hash)
+            if type != :aria && type != :class && value.is_a?(Hash)
               value.each_pair do |k, v|
                 next if k.blank? || v.nil?
 
@@ -391,6 +393,13 @@ module ActionView
       #
       #   tag.div data: { city_state: %w( Chicago IL ) }
       #   # => <div data-city-state="[&quot;Chicago&quot;,&quot;IL&quot;]"></div>
+      #
+      # In addition to <tt>:data</tt> and <tt>:aria</tt>, nest sub-attribute
+      # hashes for any keys:
+      #
+      #   tag.button "POST to /clicked", hx: { post: "/clicked", swap: :outerHTML }
+      #
+      #   # => <button hx-post="/clicked" hx-swap="outerHTML">POST to /clicked</button>
       #
       # The generated tag names and attributes are escaped by default. This can be disabled using
       # +escape+.