+++ activerecord/test/cases/validations/i18n_validation_test.rb
@@ -51,7 +51,8 @@ class I18nValidationTest < ActiveRecord::TestCase
     test "validates_uniqueness_of on generated message #{name}" do
       Topic.validates_uniqueness_of :title, validation_options
       @topic.title = unique_topic.title
-      assert_called_with(ActiveModel::Error, :generate_message, [:title, :taken, @topic, generate_message_options.merge(value: "unique!")]) do
+      error_options = generate_message_options.merge(value: "unique!", existing_id: unique_topic.id)
+      assert_called_with(ActiveModel::Error, :generate_message, [:title, :taken, @topic, error_options]) do
         @topic.valid?
         @topic.errors.messages
       end